+++ src/Randomizer/MapMusicObject.js
@@ -3,26 +3,12 @@ import tableText from '!array-loader!./tables/map_music_table.txt';
 import GridTableObject from './GridTableObject.js';
 import MapEnemyObject from './MapEnemyObject.js';
 import EnemyObject from './EnemyObject.js';
+import MusicObject from './MusicObject.js';
+
 
 class MapMusicObject extends GridTableObject {
     static shouldRandomize() {
         return this.context.specs.flags.w;
-    }
-
-    static get validMusics() {
-        if(this._validMusics !== undefined) return this._validMusics;
-        this._validMusics = new Set([2, 3, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 76, 77, 78, 80, 81, 82, 83, 84, 86, 87, 90, 91, 92, 106, 107, 108, 112, 114, 116, 117, 118, 119, 120, 121, 122, 125, 128, 129, 130, 131, 132, 136, 140, 142, 144, 146, 147, 149, 150, 151, 152, 153, 154, 159, 169, 170, 171, 173, 178, 187, 188]);
-        if(this.context.specs.flags.w >= 3) {
-            EnemyObject.every.forEach(e => this._validMusics.add(e.oldData.music));
-        }
-        return this.validMusics;
-    }
-
-    static get battleMusics() {
-        if(this._battleMusics !== undefined) return this._battleMusics;
-        if(this.context.specs.flags.w >= 3) return this.validMusics;
-        this._battleMusics = new Set(EnemyObject.every.map(e => e.oldData.music));
-        return this.battleMusics;
     }
 
     static get clearFloorFlags() {
@@ -35,14 +21,15 @@ class MapMusicObject extends GridTableObject {
         this.classReseed("mut");
         if(this.context.specs.flags.w >= 2) {
             // Battle music shuffle
-            const musics = Array.from(this.battleMusics);
+            const musics = Array.from(MusicObject.battleMusics);
             EnemyObject.every.forEach(e => e.data.music = this.context.random.choice(musics));
         }
 
 
         if(this.context.specs.flags.a) {
             // Map music shuffle - Ancient Cave mode
-            const chosenMusics = this.context.random.sample(Array.from(this.validMusics), 9);
+            const chosenMusics = MusicObject.ancientCaveMusics;
+
             MapEnemyObject.every.forEach(meo => {
                 if(meo.caveLevel === null || meo.caveLevel === undefined) return;
                 const mmo = meo.music;
@@ -66,8 +53,8 @@ class MapMusicObject extends GridTableObject {
             // Hack map music to not use lookup table
             this.context.rom.set([0x80, 0x6B], 0x6928);
 
-            const originalMusics = Array.from(this.validMusics);
-            const chosenMusics = this.context.random.shuffle(Array.from(this.validMusics));
+            const originalMusics = Array.from(MusicObject.overworldMusics);
+            const chosenMusics = this.context.random.shuffle(Array.from(MusicObject.overworldMusics));
             const assign = {};
             originalMusics.forEach((o, i) => assign[o] = chosenMusics[i]);