+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Backgrounds: New flag level 2, color-shift backgrounds. Additionally, very eye-straining backgrounds should be less common at all flag levels.
 - Game improvement patches: "Run button" changed to "Run mode"; default run speed and hold Y to slow to a walk.
 - PC sprites: Added Nessa.
+- Music: Custom music added.
 ### Bugfixes
 - Ancient Cave: Benches no longer play restoration sound effect twice.
 - Music: Remove "Teleport in" track from valid music options.